+++ packages/client/src/client.ts
@@ -272,10 +272,10 @@ export class SkipClient {
       gasAmountMultiplier = DEFAULT_GAS_MULTIPLIER,
       getFallbackGasAmount,
     } = options;
-    let gasTokenUsed: Coin | undefined;
+    let gasTokenRecord: Record<number, Coin> | undefined;
     if (validateGasBalance) {
       // check balances on chains where a tx is initiated
-      gasTokenUsed = await this.validateGasBalances({
+      gasTokenRecord = await this.validateGasBalances({
         txs,
         userAddresses,
         getOfflineSigner: options.getCosmosSigner,
@@ -287,6 +287,7 @@ export class SkipClient {
 
     for (let i = 0; i < txs.length; i++) {
       const tx = txs[i];
+      let gasTokenUsed = gasTokenRecord?.[i];
       if (!tx) {
         raise(`executeRoute error: invalid message at index ${i}`);
       }
@@ -570,16 +571,14 @@ export class SkipClient {
           );
           if (!fallbackGasAmount) {
             raise(
-              `executeRoute error: unable to estimate gas for message(s) ${
-                messages || encodedMsgs
+              `executeRoute error: unable to estimate gas for message(s) ${messages || encodedMsgs
               }`
             );
           }
           return String(fallbackGasAmount);
         }
         raise(
-          `executeRoute error: unable to estimate gas for message(s) ${
-            messages || encodedMsgs
+          `executeRoute error: unable to estimate gas for message(s) ${messages || encodedMsgs
           }`
         );
       }
@@ -647,8 +646,7 @@ export class SkipClient {
 
     if (!fee) {
       raise(
-        `executeRoute error: unable to get fee for message(s) ${
-          messages || encodedMsgs
+        `executeRoute error: unable to get fee for message(s) ${messages || encodedMsgs
         }`
       );
     }
@@ -1778,6 +1776,8 @@ export class SkipClient {
     gasAmountMultiplier?: number;
     getFallbackGasAmount?: clientTypes.GetFallbackGasAmount;
   }) {
+    // tx index -> gas token used
+    let gasTokenRecord: Record<number, Coin> = {}
     for (let i = 0; i < txs.length; i++) {
       const tx = txs[i];
       if (!tx) {
@@ -1816,8 +1816,7 @@ export class SkipClient {
           raise(
             `validateGasBalance error: invalid address for chain '${tx.cosmosTx.chainID}'`
           );
-
-        return await this.validateCosmosGasBalance({
+        const coinUsed = await this.validateCosmosGasBalance({
           client,
           signerAddress: currentAddress,
           chainID: tx.cosmosTx.chainID,
@@ -1826,8 +1825,10 @@ export class SkipClient {
           gasAmountMultiplier,
           getFallbackGasAmount,
         });
+        gasTokenRecord[i] = coinUsed;
       }
     }
+    return gasTokenRecord;
   }
 
   async validateCosmosGasBalance({
@@ -1886,11 +1887,9 @@ export class SkipClient {
           const formattedAmount =
             Number(amount.amount) / Math.pow(10, asset.decimals);
           return {
-            errMessage: `Insufficient fee token to initiate transfer on ${chainID}. Need ${formattedAmount} ${
-              asset.recommendedSymbol || amount.denom
-            }, but only have ${formattedBalance} ${
-              asset.recommendedSymbol || amount.denom
-            }.`,
+            errMessage: `Insufficient fee token to initiate transfer on ${chainID}. Need ${formattedAmount} ${asset.recommendedSymbol || amount.denom
+              }, but only have ${formattedBalance} ${asset.recommendedSymbol || amount.denom
+              }.`,
             amount,
           };
         }
@@ -1906,12 +1905,12 @@ export class SkipClient {
       if (result.length > 1) {
         throw new Error(
           result[0]?.errMessage ||
-            `Insufficient fee token to initiate transfer on ${chainID}.`
+          `Insufficient fee token to initiate transfer on ${chainID}.`
         );
       }
       throw new Error(
         result[0]?.errMessage ||
-          `Insufficient fee token to initiate transfer on ${chainID}.`
+        `Insufficient fee token to initiate transfer on ${chainID}.`
       );
     }
     return successful.amount;
@@ -1959,7 +1958,7 @@ function validateChainIDsToAffiliates(
 /**
  * @deprecated SkipRouter is deprecated please use SkipClient instead
  */
-export class SkipRouter extends SkipClient {}
+export class SkipRouter extends SkipClient { }
 
 function raise(message?: string, options?: ErrorOptions): never {
   throw new Error(message, options);