+++ packages/widget/src/hooks/use-make-wallets.tsx
@@ -99,7 +99,7 @@ export const useMakeWallets = () => {
               await client.connect();
 
               const viewService = client.service(ViewService);
-              const address = await viewService.addressByIndex({
+              const address = await viewService.ephemeralAddress({
                 addressIndex: {
                   account: penumbraWalletIndex ? penumbraWalletIndex : 0,
                 },